Adding in Special Awards
In 2021, FIRST awarded Innovation Challenge Finalist and Winner awards, but those do not appear in the official FIRST API results. We have added a generic category called "specialAwards" which allow us to include any future unrecognized award for any winning team. The variable specialAwards is structured in the same way as the return from the awards/ endpoint. Required values are:
eventCode
name (of the award)
teamNumber
schoolName
fullName

In the case it is for a specific person or people, then person needs to be populated.