Fixing iOS Viewport issues
Fixing CSS for iOS Devices, which were inheriting scale when rotating. The user experience effect was that the row of buttons at the top of the screen overflow in some scenarios.

Also updating the Team Info sheets, which merge with an export from gatool. These were specific to Infinite Recharge, and now they are generic.