24 hr time and Brand compliance
Users asked for 24 hour time display. This is now included.
We realized that our cheat sheet was out of brand compliance with *FIRST*. We have adjusted the typeface to be bold and italic everywhere it is required.
We also received feedback that the carpeted area on the cheat sheet was too dark. We corrected that as well.
Fixes #31
Fixes #30
Fixes #29