Initial commit for Scouting tool
This scouting tool will emit an Excel file and JSON (in the browser page) showing the details from a specified event. Users can supply a year and event (FIRST Event Code) in the get request, or enter a year and event code manually. The tool will retrieve the hybrid match schedule and augment it with the score details for each match. Upon export, it flattens the object into a single row for easy ingestion into analysis tools such as Tableau.