Reduced Alliance events
We have added in support for small events. These have a reduced team count, and therefore fewer than 8 Playoff Alliances.
This commit includes the logic to determine how many Playoff Alliances should be selected, and the routines to support the reduction in Playoff Alliances. It reports this number on the setup screen once an event loads.
Fixes #42