Fixing how we determine whether Alliance Selection is ready
We had been inspecting the number of matches a team played to determine whether Alliance Selection was indeed ready. The method was error prone in certain cases, because it was resetting the flag during each loop through the rows of ranks. We now count the number of unready teams, and if there is any team unready, then we do not show Alliance Selection.